Bug: fixed elf page flag bug.
The previous elf (and paging) implementation did not set the flags of a
page when the page was already present. This would mean that when a
execve was called previously (for example from userspace) and it
required a page to be read/write but it was previously readonly for code
it would cause write page faults in the code executed. The current
paging implementation now just forces flag change every time.